reference/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lad-a2a-reference"
version = "0.1.0"
description = "LAD-A2A Reference Implementation - Local Agent Discovery for A2A"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "LAD-A2A Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "zeroconf>=0.131.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.0.0",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
config = [
    "pyyaml>=6.0",
]

[project.scripts]
lad-server = "server.lad_server:main"
lad-client = "client.lad_client:main"

[tool.setuptools.packages.find]
include = ["server*", "client*", "common*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
